add propert edit, delete buttons

+++ js/default.js
@@ -35,20 +35,22 @@ $('.btn_submit_prop').click(function() {
 /**
  * Remove property.
  */
-$('.js-property-remove').confirmation({
-    onConfirm: function(event) {
-        event.preventDefault();
-        $.ajax({
-            url: '/admin/' + this.href,
-            type: 'DELETE',
-            success: function(result) {
-                if( result.length ){
-                    location.reload();
+if( typeof confirmation === "function" ) {
+    $('.js-property-remove').confirmation({
+        onConfirm: function(event) {
+            event.preventDefault();
+            $.ajax({
+                url: '/admin/' + this.href,
+                type: 'DELETE',
+                success: function(result) {
+                    if( result.length ){
+                        location.reload();
+                    }
                 }
-            }
-        });
-    }
-});
+            });
+        }
+    });
+}
 
 function login(form)
 {
